adde on change handlers to join.js cient can enter chat room given there is a chat room and name restructured folders to contain css files for each component
