+++ client/src/App.js
@@ -2,8 +2,8 @@ import React from 'react';
 
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 
-import Join from './components/Join'
-import Chat from './components/Chat'
+import Join from './components/Join/Join'
+import Chat from './components/Chat/Chat'
 
 const App = () => (
   <Router>
